--- src/ui/pages/sandbox/__init__.py
@@ -1,3 +1,0 @@
-from .sandbox_page_widget import SandboxPage
-
-__all__ = ["SandboxPage"]
